client/asserts.go: move to client/store.go
